Support mip maps generation on texture loading.
