+++ Source/HelixToolkit-Nex/HelixToolkit.Nex.Graphics.Vulkan/VulkanContext.cs
@@ -1,5 +1,3 @@
-using System.Text;
-
 namespace HelixToolkit.Nex.Graphics.Vulkan;
 
 public sealed class VulkanContextConfig()
@@ -61,6 +59,11 @@ public sealed class VulkanContextConfig()
     );
 
     public CreateSurface? OnCreateSurface = null; // custom surface creator, if not set, default surface creation will be used
+
+    /// <summary>
+    /// Force to use integrated GPU if available. This is useful for testing and debugging on systems with both integrated and discrete GPUs, but should be used with caution as it may lead to suboptimal performance. Default false.
+    /// </summary>
+    public bool ForceIntegratedGPU = false;
 };
 
 internal sealed partial class VulkanContext
@@ -727,387 +730,17 @@ internal sealed partial class VulkanContext
         return (graphicsFamily, presentFamily, computeFamily);
     }
 
-    private static unsafe VkPhysicalDevice FindBestDevice(
-        VkSurfaceKHR surface,
-        ReadOnlySpan<VkPhysicalDevice> devices,
-        byte[]? requiredLuid = null
-    )
-    {
-        List<string>? availableLuids = requiredLuid is not null ? new() : null;
-        byte* luidCopyBuffer = stackalloc byte[8];
-
-        Dictionary<VkPhysicalDeviceType, uint> bestDeviceByType = new((int)devices.Length);
-        VkPhysicalDevice selectedDevice = VkPhysicalDevice.Null;
-
-        for (uint i = 0; i < devices.Length; i++)
-        {
-            var candidate = devices[(int)i];
-
-            if (!IsDeviceSuitable(candidate, surface))
-                continue;
-
-            // When LUID filtering is active, query VkPhysicalDeviceIDProperties and skip non-matching devices
-            if (requiredLuid is not null)
-            {
-                VkPhysicalDeviceIDProperties idProps = new();
-                VkPhysicalDeviceProperties2 props2 = new() { pNext = &idProps };
-                VK.vkGetPhysicalDeviceProperties2(candidate, &props2);
-
-                byte* luidPtr = idProps.deviceLUID;
-                for (int j = 0; j < 8; j++)
-                    luidCopyBuffer[j] = luidPtr[j];
-
-                ReadOnlySpan<byte> deviceLuidSpan = new(luidCopyBuffer, 8);
-                availableLuids!.Add(Convert.ToHexString(deviceLuidSpan));
-
-                if (!idProps.deviceLUIDValid)
-                {
-                    continue;
-                }
-                if (LuidMatches(deviceLuidSpan, requiredLuid!))
-                {
-                    selectedDevice = candidate;
-                    break;
-                }
-                continue;
-            }
-
-            VK.vkGetPhysicalDeviceProperties(
-                candidate,
-                out VkPhysicalDeviceProperties checkProperties
-            );
-
-            bestDeviceByType[checkProperties.deviceType] = i;
-        }
-        if (selectedDevice.IsNotNull)
-        {
-            _logger.LogInformation(
-                $"Selected physical device based on LUID match. Device LUID: {Convert.ToHexString(requiredLuid!)}"
-            );
-            return selectedDevice;
-        }
-        if (requiredLuid is not null)
-        {
-            _logger.LogError(
-                $"Vulkan: No physical device matches the required LUID '{Convert.ToHexString(requiredLuid!)}'. "
-                    + $"Available device LUIDs: [{string.Join(", ", availableLuids!)}]"
-            );
-            return VkPhysicalDevice.Null;
-        }
-        if (bestDeviceByType.ContainsKey(VkPhysicalDeviceType.DiscreteGpu))
-        {
-            _logger.LogInformation("Selected discrete GPU.");
-            selectedDevice = devices[(int)bestDeviceByType[VkPhysicalDeviceType.DiscreteGpu]];
-        }
-        else if (bestDeviceByType.ContainsKey(VkPhysicalDeviceType.IntegratedGpu))
-        {
-            _logger.LogInformation("Selected integrated GPU.");
-            selectedDevice = devices[(int)bestDeviceByType[VkPhysicalDeviceType.IntegratedGpu]];
-        }
-        else if (bestDeviceByType.ContainsKey(VkPhysicalDeviceType.VirtualGpu))
-        {
-            _logger.LogInformation("Selected virtual GPU.");
-            selectedDevice = devices[(int)bestDeviceByType[VkPhysicalDeviceType.VirtualGpu]];
-        }
-        else if (bestDeviceByType.ContainsKey(VkPhysicalDeviceType.Cpu))
-        {
-            _logger.LogInformation("Selected CPU device.");
-            selectedDevice = devices[(int)bestDeviceByType[VkPhysicalDeviceType.Cpu]];
-        }
-
-        return selectedDevice;
-    }
-
     private unsafe ResultCode InitContext()
     {
-        // Find physical device, setup queue's and create device.
-        uint physicalDevicesCount = 0;
-        VK.vkEnumeratePhysicalDevices(_vkInstance, &physicalDevicesCount, null).CheckResult();
+        InitPhysicalDevice().CheckResult();
 
-        if (physicalDevicesCount == 0)
-        {
-            throw new Exception("Vulkan: Failed to find GPUs with Vulkan support");
-        }
+        InitDevice().CheckResult();
 
-        var physicalDevices = stackalloc VkPhysicalDevice[(int)physicalDevicesCount];
-        VK.vkEnumeratePhysicalDevices(_vkInstance, &physicalDevicesCount, physicalDevices)
-            .CheckResult();
+        InitImmediateCommands();
 
-        var luidFilterActive = Config.RequiredDeviceLuid is { Length: 8 };
-        _vkPhysicalDevice = FindBestDevice(
-            _vkSurface,
-            new ReadOnlySpan<VkPhysicalDevice>(physicalDevices, (int)physicalDevicesCount),
-            luidFilterActive ? Config.RequiredDeviceLuid : null
-        );
+        InitPipelineCache();
 
-        if (_vkPhysicalDevice.IsNull)
-        {
-            _logger.LogError("Vulkan: No suitable physical device found");
-            return ResultCode.RuntimeError;
-        }
-        VK.vkGetPhysicalDeviceProperties(
-            _vkPhysicalDevice,
-            out VkPhysicalDeviceProperties properties
-        );
-
-        if (properties.apiVersion < Config.VulkanVersion)
-        {
-            _logger.LogError("Vulkan: The physical device does not support Vulkan 1.3 or higher.");
-            return ResultCode.RuntimeError;
-        }
-        if (properties.deviceName is not null)
-        {
-            DeviceName =
-                new VkUtf8ReadOnlyString(
-                    new ReadOnlySpan<byte>(
-                        properties.deviceName,
-                        (int)VK.VK_MAX_PHYSICAL_DEVICE_NAME_SIZE
-                    )
-                ).ToString() ?? string.Empty;
-            DeviceName = DeviceName.TrimEnd('\0'); // Remove any trailing null characters
-        }
-
-        _logger.LogInformation(
-            $"""
-            Selected Graphics Card Info:
-            ---------------------------
-            Device ID: {properties.deviceID}
-            API Version: {properties.apiVersion}
-            Device Name: {DeviceName}
-            Device Type: {properties.deviceType}
-            ----------------------------
-            """
-        );
-        DeviceQueues.GraphicsQueueFamilyIndex = HxVkUtils.FindQueueFamilyIndex(
-            _vkPhysicalDevice,
-            VkQueueFlags.Graphics
-        );
-        DeviceQueues.ComputeQueueFamilyIndex = HxVkUtils.FindQueueFamilyIndex(
-            _vkPhysicalDevice,
-            VkQueueFlags.Compute
-        );
-        DeviceQueues.TransferQueueFamilyIndex = HxVkUtils.FindQueueFamilyIndex(
-            _vkPhysicalDevice,
-            VkQueueFlags.Transfer
-        );
-
-        if (DeviceQueues.GraphicsQueueFamilyIndex == DeviceQueues.INVALID)
-        {
-            _logger.LogError("VK_QUEUE_GRAPHICS_BIT is not supported");
-            return ResultCode.RuntimeError;
-        }
-        if (DeviceQueues.ComputeQueueFamilyIndex == DeviceQueues.INVALID)
-        {
-            _logger.LogError("VK_QUEUE_COMPUTE_BIT is not supported");
-            return ResultCode.RuntimeError;
-        }
-        var availableDeviceExtensions = VK.vkEnumerateDeviceExtensionProperties(_vkPhysicalDevice);
-        if (!availableDeviceExtensions.IsEmpty)
-        {
-            _supportedExtensions.Clear();
-            foreach (var ext in availableDeviceExtensions)
-            {
-                if (ext.extensionName is null)
-                {
-                    continue;
-                }
-                _supportedExtensions.Add(new VkUtf8String(ext.extensionName).ToString()!);
-            }
-        }
-
-        //var supportPresent = vkGetPhysicalDeviceWin32PresentationSupportKHR(PhysicalDevice, queueFamilies.graphicsFamily);
-        VkDeviceQueueCreateInfo* queueCreateInfos = stackalloc VkDeviceQueueCreateInfo[3];
-        float queuePriority = 1f;
-        queueCreateInfos[0] = new VkDeviceQueueCreateInfo
-        {
-            queueFamilyIndex = DeviceQueues.GraphicsQueueFamilyIndex,
-            queueCount = 1,
-            pQueuePriorities = &queuePriority, // Priority for the graphics queue
-        };
-        queueCreateInfos[1] = new VkDeviceQueueCreateInfo
-        {
-            queueFamilyIndex = DeviceQueues.ComputeQueueFamilyIndex,
-            queueCount = 1,
-            pQueuePriorities = &queuePriority, // Priority for the compute queue
-        };
-        queueCreateInfos[2] = new VkDeviceQueueCreateInfo
-        {
-            queueFamilyIndex = DeviceQueues.TransferQueueFamilyIndex,
-            queueCount = 1,
-            pQueuePriorities = &queuePriority, // Priority for the transfer queue
-        };
-
-        // Deduplicate queue families - Vulkan requires unique queue family indices
-        HashSet<uint32_t> uniqueQueueFamilies =
-        [
-            queueCreateInfos[0].queueFamilyIndex,
-            queueCreateInfos[1].queueFamilyIndex,
-        ];
-        if (DeviceQueues.TransferQueueFamilyIndex != DeviceQueues.INVALID)
-        {
-            uniqueQueueFamilies.Add(queueCreateInfos[2].queueFamilyIndex);
-        }
-        // Rebuild queueCreateInfos with only unique families
-        uint32_t numQueues = 0;
-        foreach (var familyIndex in uniqueQueueFamilies)
-        {
-            queueCreateInfos[numQueues++] = new VkDeviceQueueCreateInfo
-            {
-                queueFamilyIndex = familyIndex,
-                queueCount = 1,
-                pQueuePriorities = &queuePriority,
-            };
-        }
-
-        {
-            // Get features and properties of the physical device and create the logical device
-            // VkPhysicalDeviceMeshShaderFeaturesEXT meshShaderFeatures = new();
-            VkPhysicalDeviceVulkan13Features features1_3 = new() { pNext = null };
-            VkPhysicalDeviceVulkan12Features features1_2 = new() { pNext = &features1_3 };
-            VkPhysicalDeviceVulkan11Features features1_1 = new() { pNext = &features1_2 };
-
-            VkPhysicalDeviceFeatures2 feature_1_0 = new() { pNext = &features1_1 };
-
-            void** features_chain = &features1_2.pNext;
-
-            VkPhysicalDevice8BitStorageFeatures storage_8bit_features = default;
-            using var additionalExts = new VkStringArray(Config.ExtensionsDevice);
-            byte** pAdditionalExts = additionalExts;
-            for (int i = 0; i < additionalExts.Length; ++i)
-            {
-                _deviceExtensions.Add(new VkUtf8String(pAdditionalExts[i]));
-            }
-            if (properties.apiVersion <= VkVersion.Version_1_3)
-            {
-                if (
-                    _supportedExtensions.Contains(
-                        VK.VK_KHR_8BIT_STORAGE_EXTENSION_NAME.GetString()!
-                    )
-                )
-                {
-                    _deviceExtensions.Add(VK.VK_KHR_8BIT_STORAGE_EXTENSION_NAME);
-                    //storage_8bit_features.sType = VkStructureType.PhysicalDevice8bitStorageFeatures;
-                    *features_chain = &storage_8bit_features;
-                    features_chain = &storage_8bit_features.pNext;
-                }
-            }
-
-            // Conditionally enable VK_KHR_external_memory_win32 for DirectX interop
-            if (Config.EnableExternalMemoryWin32)
-            {
-                _deviceExtensions.Add(VK.VK_KHR_EXTERNAL_MEMORY_WIN32_EXTENSION_NAME);
-            }
-
-            VK.vkGetPhysicalDeviceFeatures2(_vkPhysicalDevice, &feature_1_0);
-            feature_1_0 = new()
-            {
-                features = DeviceFeatures.CreateFeatures10(ref feature_1_0.features),
-            };
-            features1_1 = DeviceFeatures.CreateFeatures11(ref features1_1);
-            features1_2 = DeviceFeatures.CreateFeatures12(ref features1_2);
-            features1_3 = DeviceFeatures.CreateFeatures13(ref features1_3);
-
-            feature_1_0.pNext = &features1_1;
-            features1_1.pNext = &features1_2;
-            features1_2.pNext = &features1_3;
-
-            _vkFeatures10 = feature_1_0;
-            _vkFeatures11 = features1_1;
-            _vkFeatures12 = features1_2;
-            _vkFeatures13 = features1_3;
-
-            GraphicsSettings.SupportMeshShader = _vkFeatureMeshShader.meshShader;
-            // VkPhysicalDeviceMeshShaderPropertiesEXT meshShaderProps = new();
-            VkPhysicalDeviceVulkan13Properties deviceProps1_3 = new();
-            VkPhysicalDeviceVulkan12Properties deviceProps1_2 = new() { pNext = &deviceProps1_3 };
-            VkPhysicalDeviceVulkan11Properties deviceProps1_1 = new() { pNext = &deviceProps1_2 };
-            VkPhysicalDeviceProperties2 deviceProps2 = new() { pNext = &deviceProps1_1 };
-            VK.vkGetPhysicalDeviceProperties2(_vkPhysicalDevice, &deviceProps2);
-            _vkPhysicalDeviceProperties2 = deviceProps2;
-            _vkPhysicalDeviceVulkan11Properties = deviceProps1_1;
-            _vkPhysicalDeviceVulkan12Properties = deviceProps1_2;
-            _vkPhysicalDeviceVulkan13Properties = deviceProps1_3;
-
-            VerifyRequiredDeviceExtensions();
-
-            using var deviceExtensionNames = new VkStringArray(_deviceExtensions);
-
-            VkDeviceCreateInfo deviceCreateInfo = new()
-            {
-                pNext = &feature_1_0,
-                queueCreateInfoCount = numQueues,
-                pQueueCreateInfos = queueCreateInfos,
-                enabledExtensionCount = deviceExtensionNames.Length,
-                ppEnabledExtensionNames = deviceExtensionNames,
-                pEnabledFeatures = null,
-            };
-
-            VK.vkCreateDevice(_vkPhysicalDevice, &deviceCreateInfo, null, out _vkDevice)
-                .CheckResult("Failed to create Vulkan Logical Device");
-        }
-
-        VK.vkLoadDevice(_vkDevice);
-        VK.vkGetDeviceQueue(
-            _vkDevice,
-            DeviceQueues.GraphicsQueueFamilyIndex,
-            0,
-            out GraphicsQueue.GraphicsQueue
-        );
-        VK.vkGetDeviceQueue(
-            _vkDevice,
-            DeviceQueues.ComputeQueueFamilyIndex,
-            0,
-            out GraphicsQueue.ComputeQueue
-        );
-
-        // Get the transfer queue if a dedicated one is available
-        if (DeviceQueues.TransferQueueFamilyIndex != DeviceQueues.INVALID)
-        {
-            VK.vkGetDeviceQueue(
-                _vkDevice,
-                DeviceQueues.TransferQueueFamilyIndex,
-                0,
-                out DeviceQueues.TransferQueue
-            );
-        }
-
-        if (GraphicsSettings.EnableDebug)
-        {
-            _vkDevice.SetDebugObjectName(
-                VkObjectType.Device,
-                (nuint)_vkDevice.Handle,
-                "[Vk.Device]: vkDevice"
-            );
-        }
-
-        _immediate = new VulkanImmediateCommands(
-            _vkDevice,
-            DeviceQueues.GraphicsQueueFamilyIndex,
-            HasExtDeviceFault,
-            "VkContext::immediate"
-        );
-
-        // create Vulkan pipeline cache
-        {
-            VkPipelineCacheCreateInfo ci = new()
-            {
-                flags = VkPipelineCacheCreateFlags.None,
-                initialDataSize = (uint)Config.PipelineCacheDataSize,
-                pInitialData = (void*)Config.PipelineCacheData,
-            };
-            VK.vkCreatePipelineCache(_vkDevice, &ci, null, out _pipelineCache);
-        }
-
-        if (Config.EnableVma)
-        {
-            _vma = HxVkUtils.CreateVmaAllocator(
-                _vkPhysicalDevice,
-                _vkDevice,
-                _vkInstance,
-                Config.VulkanVersion
-            );
-            HxDebug.Assert(_vma.IsNotNull);
-        }
+        InitVma();
 
         _stagingDevice = new VulkanStagingDevice(this);
 
@@ -1117,96 +750,9 @@ internal sealed partial class VulkanContext
             _transferQueue = new VulkanTransferQueue(this);
         }
 
-        // default texture
-        {
-            uint32_t pixel = 0xFF000000;
-            var result = CreateTexture(
-                new TextureDesc
-                {
-                    Format = Format.RGBA_UN8,
-                    Dimensions = new(1, 1, 1),
-                    Usage = TextureUsageBits.Sampled | TextureUsageBits.Storage,
-                    Data = (nint)(void*)&pixel,
-                    DataSize = sizeof(uint32_t), // 1x1 pixel RGBA
-                },
-                out var texture,
-                "Dummy 1x1 (black)"
-            );
-            if (result != ResultCode.Ok)
-            {
-                return result;
-            }
-            _dummyTexture = texture;
-            HxDebug.Assert(TexturesPool.Count == 1, "Dummy texture should be created successfully");
-        }
+        InitDefaultTextureSampler();
 
-        {
-            HxDebug.Assert(SamplersPool.Count == 0);
-            var result = CreateSampler(
-                new VkSamplerCreateInfo
-                {
-                    magFilter = VK.VK_FILTER_LINEAR,
-                    minFilter = VK.VK_FILTER_LINEAR,
-                    mipmapMode = VK.VK_SAMPLER_MIPMAP_MODE_LINEAR,
-                    addressModeU = VK.VK_SAMPLER_ADDRESS_MODE_REPEAT,
-                    addressModeV = VK.VK_SAMPLER_ADDRESS_MODE_REPEAT,
-                    addressModeW = VK.VK_SAMPLER_ADDRESS_MODE_REPEAT,
-                    anisotropyEnable = VK_BOOL.False,
-                    compareEnable = VK_BOOL.False,
-                    compareOp = VK.VK_COMPARE_OP_ALWAYS,
-                    minLod = 0.0f,
-                    maxLod = (Constants.MAX_MIP_LEVELS - 1),
-                    borderColor = VK.VK_BORDER_COLOR_INT_OPAQUE_BLACK,
-                    unnormalizedCoordinates = VK_BOOL.False,
-                },
-                Format.Invalid,
-                out var sampler,
-                "Sampler: default"
-            );
-            if (result != ResultCode.Ok)
-            {
-                return result;
-            }
-            _defaultSampler = new SamplerResource(this, sampler);
-            HxDebug.Assert(
-                SamplersPool.Count == 1,
-                "Default sampler should be created successfully"
-            );
-        }
-        GrowDescriptorPool(CurrentMaxTextures, CurrentMaxSamplers);
-        QuerySurfaceCapabilities();
-        {
-            var bindings = stackalloc VkDescriptorSetLayoutBinding[Constants.MAX_COLOR_ATTACHMENTS];
-            for (uint32_t i = 0; i < Constants.MAX_COLOR_ATTACHMENTS; i++)
-            {
-                bindings[i] = HxVkUtils.GetDSLBinding(
-                    i,
-                    VK.VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT,
-                    1,
-                    VK.VK_SHADER_STAGE_FRAGMENT_BIT
-                );
-            }
-
-            var limits = GetVkPhysicalDeviceProperties().limits;
-            VkDescriptorSetLayoutCreateInfo dslci = new()
-            {
-                flags = VK.VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT,
-                bindingCount = Math.Min(
-                    Constants.MAX_COLOR_ATTACHMENTS,
-                    limits.maxPerStageDescriptorInputAttachments
-                ),
-                pBindings = bindings,
-            };
-            VkDescriptorSetLayout dslInputAttachmentLayout;
-            VK.vkCreateDescriptorSetLayout(_vkDevice, &dslci, null, &dslInputAttachmentLayout)
-                .CheckResult();
-            _dslInputAttachments = dslInputAttachmentLayout;
-            _vkDevice.SetDebugObjectName(
-                VK.VK_OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT,
-                (nuint)dslInputAttachmentLayout.Handle,
-                "Descriptor Set Layout: VulkanContext::dslInputAttachments"
-            );
-        }
+        InitDescriptorSets();
         return ResultCode.Ok;
     }
 
@@ -1244,28 +790,7 @@ internal sealed partial class VulkanContext
 
     public void GenerateMipmap(in TextureHandle handle)
     {
-        if (handle.Empty)
-        {
-            return;
-        }
-
-        var tex = TexturesPool.Get(handle);
-
-        if (tex is null)
-        {
-            _logger.LogWarning("Texture {HANDLE} not found in pool", handle);
-            return;
-        }
-
-        if (tex.NumLevels <= 1)
-        {
-            return;
-        }
-
-        HxDebug.Assert(tex.ImageLayout != VK.VK_IMAGE_LAYOUT_UNDEFINED);
-        var wrapper = Immediate!.Acquire();
-        tex.GenerateMipmap(wrapper.Instance);
-        Immediate!.Submit(wrapper);
+        GenerateMipmap(handle, out var _);
     }
 
     private VkSampleCountFlags GetFramebufferMSAABitMaskVK()
